+++ internal/functions/default_functions_test.go
@@ -1,0 +1,10 @@
+package functions
+
+import "testing"
+
+func TestTitle(t *testing.T) {
+	titleCase := title("tom watkins")
+	if titleCase != "Tom Watkins" {
+		t.Fatalf("Should return in title case but returned: %v", titleCase)
+	}
+}
