+++ internal/tags/tag_checking_test.go
@@ -1,4 +1,4 @@
-package docxtpl
+package tags
 
 import (
 	"testing"
